Refactor live service and stream classes.
