+++ src/live/src/user-to-user/live-service.js
@@ -3,7 +3,6 @@ import PubNub from 'pubnub';
 
 import Client from '../../../client';
 import { KinveyRequest, RequestMethod, AuthType } from '../../../request';
-import { StreamACL } from './stream-acl';
 import { User } from '../../../entity';
 
 class LiveService {
@@ -16,8 +15,6 @@ class LiveService {
   pubnubClient;
   /** @type {string} */
   userChannelGroup;
-  /** @type {string} @private @constant */
-  streamEndpoint;
 
   /**
    * @constructor
@@ -25,7 +22,6 @@ class LiveService {
    */
   constructor(client) {
     this.client = client || Client.sharedInstance();
-    this.streamEndpoint = `/stream/${this.client.appKey}`;
   }
 
   /**
@@ -55,114 +51,12 @@ class LiveService {
   }
 
   /**
-   * @param {string} streamOwnerId
-   * @param {string} streamName
-   * @param {StreamACL} acl
-   * @returns {Promise} Response promise
-   */
-  setSubstreamACL(streamOwnerId, streamName, acl) {
-    const request = new KinveyRequest({
-      method: RequestMethod.PUT,
-      authType: AuthType.Session,
-      url: url.format({
-        protocol: this.client.apiProtocol,
-        host: this.client.apiHost,
-        pathname: `${this.streamEndpoint}/${streamName}/${streamOwnerId}`
-      }),
-      body: (acl instanceof StreamACL) ? acl.toPlainObject() : acl
-    });
-
-    return request.execute()
-      .then(response => response.data);
-  }
-
-  /**
-   * @param {string} streamName
-   * @param {string} substreamOwnerId
-   */
-  requestSubstreamPublishAccess(streamName, substreamOwnerId) {
-    const request = new KinveyRequest({
-      method: RequestMethod.POST,
-      authType: AuthType.Session,
-      url: url.format({
-        protocol: this.client.apiProtocol,
-        host: this.client.apiHost,
-        pathname: `${this.streamEndpoint}/${streamName}/${substreamOwnerId}/publish`
-      })
-    });
-
-    return request.execute()
-      .then(response => response.data);
-  }
-
-  /**
-   * @param {string} streamName
-   * @param {string} substreamOwnerId
-   */
-  requestSubstreamSubscribeAccess(streamName, substreamOwnerId) {
-    const request = new KinveyRequest({
-      method: RequestMethod.POST,
-      authType: AuthType.Session,
-      url: url.format({
-        protocol: this.client.apiProtocol,
-        host: this.client.apiHost,
-        pathname: `${this.streamEndpoint}/${streamName}/${substreamOwnerId}/subscribe`
-      }),
-      body: { deviceId: this.client.deviceId }
-    });
-
-    return request.execute()
-      .then(response => response.data);
-  }
-
-  /**
-   * @param {string} streamName
-   * @param {string} substreamOwnerId
-   */
-  unsubscribeFromSubstream(streamName, substreamOwnerId) {
-    const path = `${streamName}/${substreamOwnerId}/unsubscribe`;
-    const request = this._getStreamRequestObject(path, RequestMethod.POST, { deviceId: this.client.deviceId });
-
-    return request.execute()
-      .then(response => response.data);
-  }
-
-  /**
-   * @param {string} streamName
-   */
-  getStreamSubstreams(streamName) {
-    const request = new KinveyRequest({
-      method: RequestMethod.GET,
-      authType: AuthType.Session,
-      url: url.format({
-        protocol: this.client.apiProtocol,
-        host: this.client.apiHost,
-        pathname: `${this.streamEndpoint}/${streamName}/_substreams`
-      })
-    });
-
-    return request.execute()
-      .then(response => response.data);
-  }
-
-  /**
    * @private
    * @param {string} userId The user id
    * @returns {Promise<{publishKey: string, subscribeKey: string, authKey: string, userChannelGroup: string}>}
    */
   _makeRegisterRequest(userId) {
-    const request = new KinveyRequest({
-      method: RequestMethod.POST,
-      authType: AuthType.Session,
-      url: url.format({
-        protocol: this.client.apiProtocol,
-        host: this.client.apiHost,
-        pathname: `/user/${this.client.appKey}/${userId}/register-realtime`
-      }),
-      body: { deviceId: this.client.deviceId }
-    });
-
-    return request.execute()
+    return this._makeUserManagementRequest(`${userId}/register-realtime`)
       .then(response => response.data);
   }
 
@@ -172,41 +66,23 @@ class LiveService {
    * @returns {Promise}
    */
   _makeUnregisterRequst(userId) {
+    return this._makeUserManagementRequest(`${userId}/unregister-realtime`)
+      .then(response => response.data);
+  }
+
+  _makeUserManagementRequest(path) {
     const request = new KinveyRequest({
       method: RequestMethod.POST,
       authType: AuthType.Session,
       url: url.format({
         protocol: this.client.apiProtocol,
         host: this.client.apiHost,
-        pathname: `/user/${this.client.appKey}/${userId}/unregister-realtime`
+        pathname: `/user/${this.client.appKey}/${path}`
       }),
       body: { deviceId: this.client.deviceId }
     });
 
-    return request.execute()
-      .then(response => response.data);
-  }
-
-  /**
-   * @private
-   * @param {string} path The path after the stream/kid part
-   * @param {RequestMethod} method The request method to be used
-   * @param {Object} [body] The body of the request, if applicable
-   * @returns {Promise}
-   */
-  _getStreamRequestObject(path, method, body) {
-    const request = new KinveyRequest({
-      method: method,
-      authType: AuthType.Session,
-      url: url.format({
-        protocol: this.client.apiProtocol,
-        host: this.client.apiHost,
-        pathname: `${this.streamEndpoint}/${path}`
-      }),
-      body: body
-    });
-
-    return request;
+    return request.execute();
   }
 }
